separated ajax functions from template views, also added ajax function for retrieving user info
