+++ home/templates/index.html
@@ -51,7 +51,6 @@
                                 </div>
                                 <div class="col-12 justify-content-around mt-2">
                                     <button class="btn btn-sm btn-dark" id="reset_filters" type="button">Reset</button>
-                                    <!-- <button class="btn btn-sm btn-dark" id="apply_filters" type="button">Apply Filters</button> -->
                                 </div>
 
                             </div>
@@ -66,8 +65,6 @@
                             <a class="dropdown-item sorting-item" data-sort="data-rating">Rating</a>
                             <div class="dropdown-divider"></div>
                             <a class="dropdown-item sorting-item" data-sort="data-created">Created</a>
-                            <!-- <div class="dropdown-divider"></div> -->
-                            <!-- <a class="dropdown-item sorting-item" href="#">Updated</a> -->
                         </div>
                         <a id="sorting_order" class="btn btn-sm untouchable hidden">
                             <span class="glyphicon glyphicon-triangle-bottom untouchable"></span>
@@ -76,7 +73,6 @@
                 </div>
                 <div class="col-12 col-xl-8 col-md-7">
                     <input id="search_bar" class="fuzzy-search" placeholder="Search Keywords" size="30" aria-label="Search">
-                    <!-- <button class="btn btn-sm btn-outline-warning" type="submit">Search</button> -->
                 </div>
             </form>
         </div>
@@ -84,7 +80,6 @@
     <div class="row content-container mt-4">
         <div class="applied-filter-container mx-auto col-12"></div>
         <div id="saved_list_container" class="mx-auto">
-            <!-- <input class="fuzzy-search" /> -->
             <div class="list">
             {% include "index_helper.html" %}
             </div>
@@ -110,9 +105,28 @@ $(document).ready(function() {
 
  });
 
-var monkeyList;
+ var monkeyList;
+ var user = {anon:true, staff: false, super:false};
 
- function init_list_searcher() {
+ function get_user_info() {
+     var data = {}
+     $.ajax({
+         method: "GET",
+         url: '/ajax/user_info/',
+         data: data,
+         dataType: 'json',
+         success: set_user_type,
+     });
+ }
+
+ function set_user_type(data) {
+     user.anon = data.anon
+     user.staff = data.staff
+     user.super = data.super
+     console.log(user)
+ }
+
+ function init_list_obj() {
     monkeyList = new List('saved_list_container', {
        valueNames: [
            'name',
@@ -144,12 +158,12 @@ function tag_list_corrector(listObj) {
 }
 
  function index_handlers() {
-     init_list_searcher()
+     init_list_obj()
+     get_user_info()
      tag_list_corrector(monkeyList)
      $("#test_form").on({
          submit: e => {
              e.preventDefault()
-             // e.stopPropagation()
          }
      });
 
@@ -165,14 +179,6 @@ function tag_list_corrector(listObj) {
 
         target.toggleClass('filter-selected');
 
-        // if (target.hasClass('filter-selected')) {
-        //     target.removeClass('filter-selected')
-        // } else {
-        //      target.addClass('filter-selected')
-        // }
-        console.log(monkeyList)
-        console.log('matching items: ', monkeyList.matchingItems)
-
         var filterList = {};
         $(".filter-selected").each(function(index) {
             var category = $(this).closest(".filter-container").attr("name"),
@@ -187,20 +193,6 @@ function tag_list_corrector(listObj) {
 
         console.log('filterList: ', filterList)
 
-        // monkeyList.filter(function(item) {
-        //     var truth = false
-        //     Object.keys(filterList).forEach(function(category) {
-        //         if (filterList[category].includes(item.values()[category])) {
-        //             truth = true
-        //             return
-        //         } else {
-        //             return
-        //         }
-        //     });
-        //     return truth
-        // });
-
-        //new
         monkeyList.filter(function(item) {
             var truth = false
             Object.keys(filterList).forEach(function(category) {
@@ -260,64 +252,6 @@ function tag_list_corrector(listObj) {
          monkeyList.sort(sortBySelection, {order: sort_order});
      });
 
-     //old
-     // $("#sorting_order").on({
-     //     click: e=> {
-     //         let sort_order = $(e.target).find("span.glyphicon")
-     //         if (sort_order.hasClass("glyphicon-triangle-bottom")) {
-     //             sort_order.removeClass("glyphicon-triangle-bottom").addClass("glyphicon-triangle-top")
-     //         } else {
-     //             sort_order.removeClass("glyphicon-triangle-top").addClass("glyphicon-triangle-bottom")
-     //         }
-     //         yeet_cannon()
-     //     }
-     // })
-
-     //old
-     // $('.sorting-item').on({
-     //     click: e => {
-     //         // e.preventDefault()
-     //         let sortBySelection = $(e.target).text()
-     //         $('#sorting').text(sortBySelection)
-     //         let sort_order = $("#sorting_order")
-     //         sort_order.removeClass("hidden").removeClass("untouchable")
-     //         $('#sorting').addClass("removeCarrat")
-     //
-     //         sort_order.find("span.glyphicon").removeClass("glyphicon-triangle-top").addClass("glyphicon-triangle-bottom")
-     //         yeet_cannon()
-     //
-     //     }
-     // });
-
-     // old
-     // $("#search_bar").on({
-     //     search: e=> {
-     //         e.stopImmediatePropagation()
-     //         yeet_cannon()
-     //     }
-     // });
-
-     // $("#apply_filters, #search_bar_button").on({
-     //     click: e => {
-     //         e.preventDefault()
-     //
-     //         yeet_cannon()
-     //
-     //         $('.dropdown-menu').removeClass('show')
-     //
-     //         let tags = $.map($(".filter-selected"), x => x.innerText).join(", ")
-     //
-     //         // let filterArray = []
-     //         // let selectedFilters = $('.filter-selected')
-     //         // for (const filter of selectedFilters) {
-     //         //     filterArray.push(filter.innerText)
-     //         // }
-     //         // let displayFilters = filterArray.toString().replace(/,/g, ', ')
-     //
-     //         $('.applied-filter-container').text(`Applied Filters: ${tags}`)
-     //     }
-     // });
-
      $('#reset_filters').on({
          click: e => {
              $('.filter-selected').removeClass('filter-selected')
@@ -331,23 +265,6 @@ function tag_list_corrector(listObj) {
              e.preventDefault()
          }
      });
-
-     //old
-     // $('#reset_filters').on({
-     //     click: e => {
-     //         e.preventDefault()
-     //         e.stopImmediatePropagation()
-     //         $('.filter-selected').removeClass('filter-selected')
-     //
-     //         let tags = $.map($(".filter-selected"), x => x.innerText)
-     //
-     //         yeet_cannon()
-     //         $('.applied-filter-container').text(tags)
-     //     },
-     //     submit: e => {
-     //         e.preventDefault()
-     //     }
-     // });
  }
 
  function star_handler() {
