+++ home/urls.py
@@ -21,17 +21,19 @@ urlpatterns = [
     path('specs_and_guides', views.SpecsAndGuidesView.as_view(), name='specs_and_guides'),
     path('contact', views.ContactView.as_view(), name='contact'),
     path('success', views.SuccessView.as_view(), name='success'),
+    path('thanks', views.ThanksView.as_view(), name='thanks'),
+    path('denied', views.DeniedView.as_view(), name='denied'),
     path('ajax/delete_list/', views.delete_list, name='delete_list'),
     path('ajax/load_spec/', views.load_spec, name='load_spec'),
     path('ajax/save_rating/', views.save_rating, name='save_rating'),
     path('ajax/delete_rating/', views.delete_rating, name='delete_rating'),
     path('ajax/savedlist_info/', views.savedlist_info, name='savedlist_info'),
     path('ajax/prof_loader/', views.prof_loader, name='prof_loader'),
-    path('thanks', views.ThanksView.as_view(), name='thanks'),
-    path('denied', views.DeniedView.as_view(), name='denied'),
     path('ajax/get_item_info/', views.get_item_info, name='get_item_info'),
     path('ajax/yeet_cannon/', views.yeet_cannon, name='yeet_cannon'),
     path('ajax/flag_list/', views.flag_list, name='flag_list'),
     path('ajax/icon_list/', views.icon_list, name='icon_list'),
     path('ajax/update_icon/', views.update_icon, name='update_icon'),
+    path('ajax/user_info/', views.user_info, name='user_info'),
+
 ]
