+++ home/views/__init__.py
@@ -1,0 +1,2 @@
+from .ajax_calls import *
+from .template_views import *
